Update bee when use "bee", not at "bee pro" and detection environment
