+++ internal/app/module/beegopro/container.go
@@ -51,7 +51,6 @@ var DefaultBeegoPro = &Container{
 }
 
 func (c *Container) Run() {
-	UpdateSelf()
 	// init git refresh cache time
 	c.initTimestamp()
 	c.initUserOption()
